# burger-api-doc.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/burgers": {
      "get": {
        "operationId": "BurgerController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      },
      "post": {
        "operationId": "BurgerController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBurgerDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/burgers/{id}": {
      "get": {
        "operationId": "BurgerController_getOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "Burger API",
    "description": "Documentación de la API para Burger App",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateBurgerDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Burger Doble Queso",
            "description": "Nombre de la hamburguesa"
          },
          "description": {
            "type": "string",
            "example": "Con carne de res, doble queso, vegetales frescos y salsas",
            "description": "Descripción de la hamburguesa"
          },
          "price": {
            "type": "number",
            "example": 23000,
            "description": "Precio de la hamburguesa en pesos colombianos"
          },
          "imageUrl": {
            "type": "string",
            "example": "https://i.imgur.com/xXZPZyZ.png",
            "description": "URL de la imagen de la hamburguesa"
          }
        },
        "required": [
          "name",
          "description",
          "price",
          "imageUrl"
        ]
      },
...
    }
  }
}